Streams rewrite
This is a major change.  To `read()`, the poller and the bindings layer, there isn’t a great way to do this piecemeal. The goal here is to make bindings a plugin layer, our c++ bindings isolated and tested, and make serialport a valid streams object.

 - Move all bindings into platform specific files
 - Provide mock bindings as a tested first class bindings layer
 - Separate out the binary bindings from the SerialPort JS so it can be required without bindings
 - Make Bindings manage .isOpen and FD as file descriptors are an implementation detail
 - Make Bindings an instantiated object so they can isolate their own state, have setup code, etc.
 - Since all integration tests now require an arduino ditch the "integration-light" tests and test with both native and mock bindings

TODO

 - Figure out how we’re going to do _read
 - ensure docs are up to date with sp methods
 - ensure bindings are documented
 - redo parsers as transform streams
 - update parser documentation
 - performance test
 - 🎉

We're currently providing `push` to the binding constructor and asking the binding to provide a `_read` implementation. Considering a `read(function(data){});` interface instead. 

The c++ bindings use .open to create some global state around write queues and read pollers. Now that the stream will only have a single call to write out at any time this is probably not necessary. For Reads on the other hands it might absolutely be necessary. In any case, hiding this state in a “global scope” is bad practice.